+++ packages/react-components/src/components/steamship-chat-elements.tsx
@@ -1,47 +1,70 @@
 import { ReactNode } from "react";
-import { Skeleton } from "../ui";
+import { Skeleton, cn } from "../ui";
 import { BotIcon, UserIcon } from "lucide-react";
+import clsx from "clsx";
 
 export const SteamshipChatMessageContainer = ({
   children,
+  className,
 }: {
   children: ReactNode;
+  className?: string;
 }) => (
-  <div className="steamship-border steamship-border-foreground/10 steamship-px-2 steamship-py-4 steamship-rounded-md steamship-mb-4 steamship-grid steamship-grid-cols-12">
+  <div
+    className={cn(
+      "steamship-border steamship-border-foreground/10 steamship-px-2 steamship-py-4 steamship-rounded-md steamship-mb-4 steamship-grid steamship-grid-cols-12",
+      className
+    )}
+  >
     {children}
   </div>
 );
 
 export const SteamshipChatUserContainer = ({
   children,
+  className,
 }: {
   children: ReactNode;
+  className?: string;
 }) => (
-  <div className="steamship-col-span-1 steamship-flex steamship-justify-center">
+  <div
+    className={cn(
+      "steamship-col-span-1 steamship-flex steamship-justify-center",
+      className
+    )}
+  >
     {children}
   </div>
 );
 
 export const SteamshipChatUser = ({
   role,
+  className,
+  iconClassName,
 }: {
   role: "function" | "user" | "system" | "assistant";
+  className?: string;
+  iconClassName?: string;
 }) => (
-  <SteamshipChatUserContainer>
+  <SteamshipChatUserContainer className={className}>
     {role === "user" ? (
-      <UserIcon className="steamship-h-6 steamship-w-6" />
+      <UserIcon className={cn("steamship-h-6 steamship-w-6", iconClassName)} />
     ) : (
-      <BotIcon className="steamship-h-6 steamship-w-6" />
+      <BotIcon className={cn("steamship-h-6 steamship-w-6", iconClassName)} />
     )}
   </SteamshipChatUserContainer>
 );
 
 export const SteamshipChatMessageContentsContainer = ({
   children,
+  className,
 }: {
   children: ReactNode;
+  className?: string;
 }) => (
-  <div className="steamship-space-y-2 steamship-col-span-11">{children}</div>
+  <div className={cn("steamship-space-y-2 steamship-col-span-11", className)}>
+    {children}
+  </div>
 );
 
 export const SteamshipChatLoadingMessage = () => (